First Try at Shiny App
This is simple rough draft for a shiny app that will allow users to select independent values they want and and run the functions from this package, using a base R data set.

Right now it only works when you select 1 variable; you get a non-conformable objects error if you select more. But I'm working on