Completed Shiny App
Ran into too many roadblocks with the version that let the user pick which independent variables they wanted to use.

This version has the user type in a "vector" for x and y. Once user presses the button the outputs for this package are printed, each separated in its own tab.